requirements.txt
# # requirements.txt
# Flask==2.3.3
# Flask-CORS==4.0.0
# python-dotenv==1.0.0
# requests==2.31.0
# PyJWT==2.8.0
# python-dateutil==2.8.2
# pydantic==2.4.2
# ujson==5.8.0
# gunicorn==21.2.0
# Flask-SQLAlchemy==3.0.5
# psycopg2-binary==2.9.7        # For PostgreSQL connection
# supabase==1.1.0               # Supabase client


# # Optional for local AI
# ollama==0.1.6
# transformers==4.35.0
# torch==2.1.0
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9        # For PostgreSQL connection
supabase==2.12.0              # Supabase client
google-genai==1.3.0    # Required for Gemini integration
pydantic==2.5.2        # Required for Structured Outputs